Add Room lib to handle offline working
A SQLite database has been chosen since it allows an easy and quick implementation and an optimal integration with kotlin and Android arch components.
Data classes, in this case RedditUiImage, can easily be inserted or queried from database.
With these changes the repository will use both network and db to deliver results to the UI.
Specifically the repository uses this logic:
- A query is performed to the repository
- First of all the repository performs the network call.
- If result is empty then the db is queried and results are returned to UI
- Otherwise network result is saved to db and returned to the UI
In this way offline queries that returns some results are one successfully completed previously.

A change in dependencies construction was necessary due to context needed for db creation.

Image caching duty has been left to Glide lib, so if an image has been previously open it can be shown later, also when offline, since Glide by default uses a memory and a disk cache.
An improvement here can be to download result images in advance, so we can show also previously unseen images, but it can be expensive if we look at generated network traffic.
Maybe we can add an option in a settings screen later